ENH Make reference reading faster
This was not a problem for databases with few large sequences, but when
using many sequences, small inefficiencies in header parsing added up

There is a slight loss of features here as duplicated names will only be
checked later and it is not difficult to report the line number.

Parsing the greengenes database (2022.10.seqs.fna, with 21,090,001
sequences) goes from 19-20s to 7s.